refactor(orders): restructure order detail page for clarity
Replaces equal-weight cards with a clearer hierarchy:

- Identity strip dedupes header/fulfillment: status badges, slim pickup
  progress, one primary action, overflow menu for cancel/refund/copy-link.
- Single line-items "document" card: service rows -> right-side sheet,
  products, then a receipt-foot money summary below all line items.
- Service detail lives in a sheet, read live from the order cache so
  in-sheet status changes reflect without a frozen prop.
- Responsive desktop/tablet/mobile; long mono order/service codes wrap
  instead of end-truncating (distinguishing tail stays visible).
- Extract OrderReasonCallout for the shared cancel/refund block; split
  getOrderServiceItemDetails to drop a magic-string compare; remove dead
  Array.isArray guards on type-guaranteed arrays.
